Add ci pipeline for notebooks-beta-qa
  - Trigger deployment for pebbles-deploy master branch
  - Copy vault password to the file path with proper selinux configs
  - Add possibility to copy vault password depending on the cluster
  - Skip formatting docker device if already present

+++ container-src/pebbles-deployer/init_env.bash
@@ -18,14 +18,20 @@ if [[ ! -e /dev/shm/secret/vaultpass ]]; then
     mkdir -p /dev/shm/secret/
     touch /dev/shm/secret/vaultpass
     chmod 600 /dev/shm/secret/vaultpass
-    if [[ -z ${VAULT_PASS} ]]; then
-        read -s -p "vault password: " VAULT_PASS
-        echo
-    fi
-    echo $VAULT_PASS > /dev/shm/secret/vaultpass
-    echo "Wrote vault password to /dev/shm/secret/vaultpass"
-    unset VAULT_PASS
+    if [[ -e /run/pebbles/secret/vaultpass_$env_name ]]; then
+        cp /run/pebbles/secret/vaultpass_$env_name /dev/shm/secret/vaultpass
+        echo "vaultpass is copied for $env_name"
+    else
+       if [[ -z ${VAULT_PASS} ]]; then
+           read -s -p "vault password: " VAULT_PASS
+       fi
+       echo $VAULT_PASS > /dev/shm/secret/vaultpass
+       echo "Wrote vault password to /dev/shm/secret/vaultpass"
+       unset VAULT_PASS
+    fi 
+
 fi
+
 
 export ANSIBLE_INVENTORY=$HOME/pebbles-environments/$env_name
 echo "ANSIBLE_INVENTORY set to $ANSIBLE_INVENTORY"
